pyproject.toml
[project]
name = "semesterproject"
version = "0.1.0"
description = "Image Scaling Prompt Injection Attacks Project"
readme = "steps.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "Flask>=3.1.0,<4.0.0",
    "Pillow>=10.0.0",
    "numpy>=2.1.0,<3.0.0",
    "bleach>=6.0.0,<7.0.0",
    "MarkupSafe>=2.1.0,<3.0.0",
    "opencv-python>=4.10.0,<5.0.0",
    "tensorflow>=2.20.0,<3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
